refactor: Extract a measure helper
